Python 3.10 and Numpy 1.21 Support

+++ bqskit/exec/runners/sim.py
@@ -13,5 +13,5 @@ class SimulationRunner(CircuitRunner):
     def run(self, circuit: Circuit) -> RunnerResults:
         """Execute the circuit, see CircuitRunner.run for more info."""
         state = np.reshape(circuit.get_unitary()[:, 0], (-1))
-        probs = np.square(np.abs(state))
-        return RunnerResults(circuit.num_qudits, circuit.radixes, probs)
+        probs = np.square(np.abs(state), dtype=np.float64)
+        return RunnerResults(circuit.num_qudits, circuit.radixes, probs)  # type: ignore # noqa
